# [FIX] sap.m.SuggestionsPopover: Improve highlightSuggestionItems performance ~300%