@mauriciolauffer 2020-04-17
Clean fork based on PR https://github.com/SAP/openui5/pull/2867

![image](https://user-images.githubusercontent.com/443888/79540819-1c051400-80cc-11ea-8588-63e7272317df.png)