@mauriciolauffer 2020-04-17
Same Travis-CI error as in PR https://github.com/SAP/openui5/pull/2867 which means the error had nothing to do with the PR or its content.

Have a look at the Travis logs...