@Shtilianova 2020-04-21
Hello @mauriciolauffer ,
Thank you for sharing this finding. I've created an internal incident 2070154734. The status of the issue will be updated here in GitHub.
Regards, Diana