@mauriciolauffer 2020-04-23
@d3xter666 thanks for spotting it. Either just grouping .innerHTML or just using .textContent produces pretty much the same result in terms of performance improvements. Both together reduces processing time and layout thrashing even further!

![image](https://user-images.githubusercontent.com/443888/80047443-ccea3380-8550-11ea-882c-5f83927f9240.png)