@matz3 2020-04-24
@mauriciolauffer could you please rebase your commits onto the latest SAP/openui5 master?

We've recently updated the .travis.yml to use Node 10 (https://github.com/SAP/openui5/commit/8c298fbcde977bfc67ada69bac2a2148142b3652), which should solve the Travis CI build issue.