@niyap 2020-04-29
Hello @mauriciolauffer,

Thank you for sharing this finding!
Your pull request has been submitted.

Kind Regards,
Niya